Support explicit cluster bus port
Support an explicit cluster bus port in addition to the implicit cluster
bus port (of service port + 10000), including upgrade path to using this
feature. Specifically:

- Add configuration option, cluster-port, for specifying cluster port
- Add optional cluster port to MEET command (if omitted, assume implicit
  port is used by the host being met).
- Add optional cluster port to persisted nodes.conf and response to
  CLUSTER NODES command (omitted if the cluster bus port, whether
  explicit or implicit, is equal to the implicit cluster port).

If the implicit port is used (even if expressed explicitly), the API
remains unchanged (nodes.conf and the response to CLUSTER NODES will be
the same), allowing existing clients (that don't support the additional
port) to continue to function.

A cluster may comprise a mix of both nodes that support an explicit port
and those that don't, provided the explicit port feature is not used.
This allows for an in-place upgrade path from a cluster without support
to a cluster that uses an explicit port.